separate service for customerClaimService and add @Transactional for all method that connect to database
