+++ src/main/java/com/app/smartdrive/api/services/customer/impl/CustomerRequestServiceImpl.java
@@ -18,17 +18,14 @@ import com.app.smartdrive.api.repositories.customer.CustomerClaimRepository;
 import com.app.smartdrive.api.repositories.customer.CustomerInscDocRepository;
 import com.app.smartdrive.api.repositories.customer.CustomerInscExtendRepository;
 import com.app.smartdrive.api.repositories.master.*;
-import com.app.smartdrive.api.services.customer.CustomerInscAssetsService;
-import com.app.smartdrive.api.services.customer.CustomerInscDocService;
-import com.app.smartdrive.api.services.customer.CustomerInscExtendService;
-import com.app.smartdrive.api.services.customer.CustomerRequestService;
+import com.app.smartdrive.api.services.customer.*;
 import com.app.smartdrive.api.services.users.BusinessEntityService;
 import com.app.smartdrive.api.services.users.UserService;
-import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 
@@ -47,8 +44,6 @@ import lombok.RequiredArgsConstructor;
 public class CustomerRequestServiceImpl implements CustomerRequestService {
     private final CustomerRequestRepository customerRequestRepository;
 
-    private final BusinessEntityRepository businessEntityRepo;
-
     private final CarsRepository carsRepository;
 
     private final IntyRepository intyRepository;
@@ -58,12 +53,6 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
     private final UserRepository userRepository;
 
     private final TemiRepository temiRepository;
-
-    private final CustomerInscExtendRepository cuexRepository;
-
-    private final CustomerInscDocRepository cadocRepository;
-
-    private final CustomerClaimRepository customerClaimRepository;
 
     private final BusinessEntityService businessEntityService;
 
@@ -75,23 +64,31 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
 
     private final CustomerInscExtendService customerInscExtendService;
 
+    private final CustomerClaimService customerClaimService;
 
+
+    @Transactional(readOnly = true)
+    @Override
     public List<CustomerRequest> get(){
         return this.customerRequestRepository.findAll();
     }
 
+    @Transactional(readOnly = true)
+    @Override
     public Page<CustomerResponseDTO> getPaging(Pageable pageable){
         Page<CustomerRequest> pageCustomerRequest = this.customerRequestRepository.findAll(pageable);
         Page<CustomerResponseDTO> pageCustomerResponseDTO = pageCustomerRequest.map(new Function<CustomerRequest, CustomerResponseDTO>() {
             @Override
             public CustomerResponseDTO apply(CustomerRequest customerRequest) {
-                return convert(customerRequest);
+                return TransactionMapper.mapEntityToDto(customerRequest, CustomerResponseDTO.class);
             }
         });
 
         return pageCustomerResponseDTO;
     }
 
+    @Transactional(readOnly = true)
+    @Override
     public CustomerResponseDTO getCustomerRequestById(Long creqEntityId){
         CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(creqEntityId)
                 .orElseThrow(() -> new EntityNotFoundException("Customer Request dengan id ${creqEntityId} tidak ditemukan")
@@ -99,7 +96,9 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
 
         return TransactionMapper.mapEntityToDto(existCustomerRequest, CustomerResponseDTO.class);
     }
+
     @Transactional
+    @Override
     public CustomerResponseDTO create(@Valid CustomerRequestDTO customerRequestDTO, MultipartFile[] files) throws Exception {
         // prep
         CiasDTO ciasDTO = customerRequestDTO.getCiasDTO();
@@ -128,7 +127,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         cias.setCiasTotalPremi(premi);
         cias.setCustomerInscExtend(ciasCuexs);
 
-        CustomerClaim newClaim = this.createNewClaim(newCustomerRequest);
+        CustomerClaim newClaim = this.customerClaimService.createNewClaim(newCustomerRequest);
 
         // set and save
         newCustomerRequest.setCustomerClaim(newClaim);
@@ -138,6 +137,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         return TransactionMapper.mapEntityToDto(savedCreq, CustomerResponseDTO.class);
     }
 
+    @Override
     public CustomerResponseDTO convert(CustomerRequest customerRequest){
         CustomerInscAssets cias = customerRequest.getCustomerInscAssets();
         List<CustomerInscExtend> cuexList = cias.getCustomerInscExtend();
@@ -321,6 +321,8 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         return new CustomerResponseDTO();
     }
 
+    @Transactional(readOnly = true)
+    @Override
     public Double getPremiPrice(String insuraceType, String carBrand, Long zonesId, Double currentPrice, List<CustomerInscExtend> cuexs){
         TemplateInsurancePremi temiMain = this.temiRepository.findByTemiZonesIdAndTemiIntyNameAndTemiCateId(zonesId, insuraceType, 1L);
 
@@ -340,6 +342,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         return totalPremi;
     }
 
+    @Transactional(readOnly = true)
     @Override
     public Page<CustomerResponseDTO> getPagingUserCustomerRequests(Long custId, Pageable paging, String type, String status) {
         User user = this.userRepository.findById(custId).get();
@@ -357,13 +360,14 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         Page<CustomerResponseDTO> pageCustomerResponseDTO = pageCustomerRequest.map(new Function<CustomerRequest, CustomerResponseDTO>() {
             @Override
             public CustomerResponseDTO apply(CustomerRequest customerRequest) {
-                return convert(customerRequest);
+                return TransactionMapper.mapEntityToDto(customerRequest, CustomerResponseDTO.class);
             }
         });
 
         return pageCustomerResponseDTO;
     }
 
+    @Transactional
     @Override
     public CustomerResponseDTO updateCustomerRequest(Long creqEntityId, UpdateCustomerRequestDTO updateCustomerRequestDTO, MultipartFile[] files) throws Exception {
         CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(creqEntityId)
@@ -406,7 +410,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         existCustomerRequest.setCreqModifiedDate(LocalDateTime.now());
 
         CustomerRequest savedCustomerRequest = this.customerRequestRepository.save(existCustomerRequest);
-        return this.convert(savedCustomerRequest);
+        return TransactionMapper.mapEntityToDto(savedCustomerRequest, CustomerResponseDTO.class);
     }
 
     @Transactional
@@ -419,53 +423,12 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         this.customerRequestRepository.delete(existCustomerRequest);
     }
 
-
-    @Override
-    public CustomerClaim createNewClaim(CustomerRequest customerRequest) {
-        CustomerClaim newCustomerClaim = CustomerClaim.builder()
-                .cuclEventPrice(0.0)
-                .cuclSubtotal(0.0)
-                .cuclEvents(0)
-                .cuclCreqEntityid(customerRequest.getCreqEntityId())
-                .customerRequest(customerRequest)
-                .build();
-
-        return newCustomerClaim;
-    }
-
-
-    @Override
-    public ClaimResponseDTO getCustomerClaimById(Long cuclCreqEntityId) {
-        CustomerClaim existCustomerClaim = this.customerClaimRepository.findById(cuclCreqEntityId).orElseThrow(
-                () -> new EntityNotFoundException("Customer Claim dengan id " + cuclCreqEntityId + " tidak ditemukan")
-        );
-
-
-       return ClaimResponseDTO.builder()
-               .cuclCreqEntityId(existCustomerClaim.getCuclCreqEntityid())
-               .cuclCreateDate(existCustomerClaim.getCuclCreateDate())
-               .cuclReason(existCustomerClaim.getCuclReason())
-               .cuclEventPrice(existCustomerClaim.getCuclEventPrice())
-               .cuclSubtotal(existCustomerClaim.getCuclSubtotal())
-               .build();
-
-    }
-
-    @Override
-    public void deleteCustomerClaim(Long cuclCreqEntityId) {
-        CustomerClaim existCustomerClaim = this.customerClaimRepository.findById(cuclCreqEntityId).orElseThrow(
-                () -> new EntityNotFoundException("Customer Claim dengan id " + cuclCreqEntityId + " tidak ditemukan")
-        );
-
-        this.customerClaimRepository.delete(existCustomerClaim);
-    }
-
     @Override
     public CustomerRequest createCustomerRequest(
             BusinessEntity newEntity,
             User customer,
             Long entityId
-            ){
+    ){
 
         return CustomerRequest.builder()
                 .businessEntity(newEntity)
@@ -477,10 +440,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
                 .build();
     }
 
-
-
-
-
+    @Transactional
     @Override
     public void changeRequestTypeToPolis(CustomerRequestTypeDTO customerRequestTypeDTO) {
         CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(customerRequestTypeDTO.getCreqEntityId()).orElseThrow(
@@ -492,41 +452,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         this.customerRequestRepository.save(existCustomerRequest);
     }
 
-    @Override
-    public CustomerResponseDTO updateCustomerClaim(ClaimRequestDTO claimRequestDTO) {
-        CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(claimRequestDTO.getCreqEntityId()).orElseThrow(
-                () -> new EntityNotFoundException("Customer Claim dengan id " + claimRequestDTO.getCreqEntityId() + " tidak ditemukan")
-        );
-
-        existCustomerRequest.setCreqType(EnumCustomer.CreqType.CLAIM);
-        existCustomerRequest.setCreqModifiedDate(LocalDateTime.now());
-
-        CustomerClaim existCustomerClaim = existCustomerRequest.getCustomerClaim();
-
-        LocalDateTime cuclCreateDate = existCustomerClaim.getCuclCreateDate();
-
-        if(Objects.isNull(cuclCreateDate)){
-            existCustomerClaim.setCuclCreateDate(LocalDateTime.now());
-        }
-
-
-        Double cuclEventPrice = existCustomerClaim.getCuclEventPrice();
-        cuclEventPrice += claimRequestDTO.getCuclEventPrice();
-
-        Double cuclSubtotal = existCustomerClaim.getCuclSubtotal();
-        cuclSubtotal += claimRequestDTO.getCuclSubtotal();
-
-        int cuclEvents = existCustomerClaim.getCuclEvents();
-        cuclEvents += 1;
-
-        existCustomerClaim.setCuclEventPrice(cuclEventPrice);
-        existCustomerClaim.setCuclSubtotal(cuclSubtotal);
-        existCustomerClaim.setCuclEvents(cuclEvents);
-
-        CustomerRequest savedCustomerRequest = this.customerRequestRepository.save(existCustomerClaim.getCustomerRequest());
-        return this.convert(savedCustomerRequest);
-    }
-
+    @Transactional
     @Override
     public void changeRequestTypeToClaim(CustomerRequestTypeDTO customerRequestTypeDTO) {
         CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(customerRequestTypeDTO.getCreqEntityId()).orElseThrow(
@@ -538,7 +464,7 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         this.customerRequestRepository.save(existCustomerRequest);
     }
 
-
+    @Transactional
     @Override
     public CustomerResponseDTO closePolis(CloseRequestDTO closeRequestDTO) {
         CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(closeRequestDTO.getCreqEntityId()).orElseThrow(
@@ -552,9 +478,10 @@ public class CustomerRequestServiceImpl implements CustomerRequestService {
         customerClaim.setCuclCreateDate(LocalDateTime.now());
 
         CustomerRequest savedCustomerRequest = this.customerRequestRepository.save(existCustomerRequest);
-        return this.convert(savedCustomerRequest);
+        return TransactionMapper.mapEntityToDto(savedCustomerRequest, CustomerResponseDTO.class);
     }
 
+    @Transactional
     @Override
     public void changeRequestTypeToClose(CustomerRequestTypeDTO customerRequestTypeDTO) {
         CustomerRequest existCustomerRequest = this.customerRequestRepository.findById(customerRequestTypeDTO.getCreqEntityId()).orElseThrow(
